updated user policy and test class

+++ app/Http/Controllers/UserController.php
@@ -22,7 +22,11 @@ class UserController extends Controller
     // creates a new user
     public function store(Request $request)
     {
-        Gate::authorize('create', auth()->user());
+        $can_store = Gate::inspect('create', auth()->user());
+        if(!$can_store->allowed()) {
+            return redirect()->back()->with('error', 'Invalid permission');
+        } 
+
         $validated_data = $request->validate([
             'name' => 'required',
             'email' => 'required|email|unique:users',
@@ -53,6 +57,11 @@ class UserController extends Controller
     // updates the user information
     public function update(Request $request, User $user)
     {
+        $can_store = Gate::inspect('update', auth()->user());
+        if(!$can_store->allowed()) {
+            return redirect()->back()->with('error', 'Invalid permission');
+        }
+
         $validated_user = $request->validate([
             'name' => 'required',
             'email' => 'required|email',
@@ -73,6 +82,11 @@ class UserController extends Controller
 
     public function destroy(User $user)
     {
+        $can_store = Gate::inspect('delete', auth()->user());
+        if(!$can_store->allowed()) {
+            return redirect()->back()->with('error', 'Invalid permission');
+        } 
+
         $user->roles()->detach();
         $user->delete();
